3dhub-website-main/backend_django/requirements.txt
asgiref==3.7.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.2
Django==4.2.1
django-appconf==1.0.5
django-cors-headers==4.0.0
django-hashedfilenamestorage==2.4.1
django-imagekit==4.1.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
idna==3.4
pilkit==2.0
Pillow==9.5.0
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.7.0
pytz==2023.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.2
tzdata==2023.3
Unidecode==1.3.6
urllib3==2.0.3
